refactor: api에서 받아온 영화 정보를 정의하는 타입을 별도 파일로 분리
